logistic_chatbot/requirements.txt
streamlit==1.32.0
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
groq==0.5.0
protobuf==3.20.0